+++ services/whos_on_duty.rb
@@ -1,6 +1,7 @@
 require 'csv'
 require_relative 'ir_pagerduty'
 require_relative '../lib/support_rota_doc'
+require_relative '../lib/builder'
 
 module WhosOnDuty
   extend self
@@ -8,13 +9,9 @@ module WhosOnDuty
   def list
     source.fetch_data
 
-    webops   = parse_webops(source.webops(:current))
-    devs     = parse_devs(source.devs(:current), source.devs(:next))
-    managers = pagerduty_managers
-
-    # Fall back to google doc if no IRM in PagerDuty
-    managers      = managers.any? ? managers : source.duty_managers(:current)
-    duty_managers = parse_duty_managers(managers)
+    webops        = Builder::Webop.hash(source.webops(:current))
+    devs          = Builder::Dev.hash(source.devs(:current), source.devs(:next))
+    duty_managers = Builder::Manager.hash(fetch_managers)
 
     webops + devs + duty_managers
 
@@ -36,60 +33,9 @@ module WhosOnDuty
     pagerduty.fetch_todays_schedules_users(SupportApp.pager_duty_irm_schedule_id)
   end
 
-  def build_row(name, rule, has_phone, contact_methods = [])
-    {
-      name:            name,
-      rule:            rule,
-      has_phone:       has_phone,
-      contact_methods: contact_methods
-    }
-  end
-
-  def parse_webops(webops)
-    webops.map { |webop| build_row(webop, 'webop', true) }
-  end
-
-  def parse_devs(current_devs, next_devs)
-    #
-    # dev with phone if:
-    # - is the only dev today
-    #  OR
-    # - is the one who has just joined (i.e on duty next week)
-    #
-    current_devs.map do |dev|
-      has_phone = next_devs.include?(dev) || current_devs.count == 1
-      build_row(dev, 'dev', has_phone)
-    end
-  end
-
-  def parse_duty_managers(managers)
-    managers.map do |manager|
-      build_row(
-        manager['name'], 
-        'duty_manager', 
-        false, 
-        manager['contact_methods'].map(&method(:build_contact_method_row))
-      )
-    end
-  end
-
-  def build_contact_method_row(contact_method)
-    {
-      type:    contact_method['type'],
-      address: format_contact_method_address(contact_method),
-      label:   contact_method['label'],
-    }
-  end
-
-  def format_contact_method_address(contact_method)
-    address = contact_method['address']
-
-    if ['phone', 'SMS'].include? contact_method['type']
-      address = 
-      "(00) #{contact_method['country_code']} " + 
-      "#{contact_method['phone_number'].reverse.gsub(/.{4}(?=.)/, '\0 ').reverse}"
-    end
-
-    address
+  # Fall back to google doc if no IRM in PagerDuty
+  def fetch_managers
+    managers = pagerduty_managers
+    managers = managers.any? ? managers : source.duty_managers(:current)
   end
 end
